Add schedule 2 species details to ROPs playback
It was missing before. Now it's not missing.
